Permet la suppression même si la société est dans un état incohérent

+++ project/plugins/acVinSocietePlugin/lib/model/Societe/Societe.class.php
@@ -437,7 +437,9 @@ class Societe extends BaseSociete implements InterfaceCompteGenerique {
 
     public function delete() {
       foreach($this->getComptesAndEtablissements() as $id => $obj) {
+        if ($obj) {
           $obj->delete();
+        }
       }
       return parent::delete();
     }
